sessions: split harness lifecycle from provider event mapping
