+++ rust/loopflow/src/lfd/sessions/harness/claude.rs
@@ -1,19 +1,17 @@
-use std::collections::HashMap;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
-use serde_json::Value;
 use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::process::{Child, Command};
 use tokio::sync::broadcast;
 use tokio::task::JoinHandle;
 
-use crate::lfd::sessions::harness::{SessionHarness, SessionHarnessError};
-use crate::lfd::sessions::types::{
-    FileEdit, ItemStatus, SessionConfig, SessionEvent, SessionItem, TurnStatus,
-};
+use crate::lfd::sessions::harness::claude_mapping::ReaderState;
+use crate::lfd::sessions::harness::common::{spawn_stderr_logger, TurnInProgressGuard};
+use crate::lfd::sessions::harness::{claude_mapping, SessionHarness, SessionHarnessError};
+use crate::lfd::sessions::types::{SessionConfig, SessionEvent, TurnStatus};
 
 pub struct ClaudeHarness {
     events: broadcast::Sender<SessionEvent>,
@@ -29,75 +27,6 @@ pub struct ClaudeHarness {
 impl std::fmt::Debug for ClaudeHarness {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("ClaudeHarness").finish()
-    }
-}
-
-/// Map Claude tool name to a SessionItem category.
-fn infer_item(tool_name: &str, tool_use_id: &str, input: Option<Value>) -> SessionItem {
-    build_item(tool_name, tool_use_id, input, ItemStatus::InProgress, None)
-}
-
-fn build_item(
-    tool_name: &str,
-    tool_use_id: &str,
-    input: Option<Value>,
-    status: ItemStatus,
-    output: Option<String>,
-) -> SessionItem {
-    match tool_name {
-        "Bash" => SessionItem::Command {
-            id: tool_use_id.to_string(),
-            command: command_from_input(input.as_ref()),
-            cwd: String::new(),
-            status,
-            output,
-            exit_code: None,
-            duration_ms: None,
-        },
-        "Edit" | "Write" | "NotebookEdit" => SessionItem::File {
-            id: tool_use_id.to_string(),
-            changes: file_changes_from_input(tool_name, input.as_ref()),
-            status,
-        },
-        _ => SessionItem::Tool {
-            id: tool_use_id.to_string(),
-            name: tool_name.to_string(),
-            status,
-            input,
-            output,
-        },
-    }
-}
-
-fn command_from_input(input: Option<&Value>) -> Vec<String> {
-    let command = input
-        .and_then(|value| value.get("command"))
-        .and_then(Value::as_str)
-        .unwrap_or_default();
-    if command.is_empty() {
-        Vec::new()
-    } else {
-        vec![command.to_string()]
-    }
-}
-
-fn file_changes_from_input(tool_name: &str, input: Option<&Value>) -> Vec<FileEdit> {
-    let path = input
-        .and_then(|value| {
-            value
-                .get("file_path")
-                .or_else(|| value.get("notebook_path"))
-        })
-        .and_then(Value::as_str)
-        .unwrap_or_default();
-    if path.is_empty() {
-        Vec::new()
-    } else {
-        vec![FileEdit {
-            path: path.to_string(),
-            kind: Some(tool_name.to_lowercase()),
-            diff: None,
-        }]
     }
 }
 
@@ -138,101 +67,6 @@ fn build_args(content: &str, config: &SessionConfig, resume_id: Option<&str>) ->
     args
 }
 
-/// Reader-local state for tracking in-flight content blocks.
-#[derive(Debug, Default)]
-struct ToolUseState {
-    id: String,
-    name: String,
-    input_json: String,
-    input: Option<Value>,
-}
-
-impl ToolUseState {
-    fn parsed_input(&self) -> Option<Value> {
-        if self.input_json.is_empty() {
-            return self.input.clone();
-        }
-
-        serde_json::from_str(&self.input_json)
-            .ok()
-            .or_else(|| self.input.clone())
-    }
-}
-
-#[derive(Debug, Default)]
-struct ReaderState {
-    /// Map block index -> tool use state for streaming deltas.
-    tool_blocks: HashMap<usize, ToolUseState>,
-    /// Map tool_use_id -> block index for direct completion lookup.
-    tool_indexes: HashMap<String, usize>,
-}
-
-impl ReaderState {
-    fn track_tool(
-        &mut self,
-        index: usize,
-        tool_id: &str,
-        tool_name: &str,
-        input: Option<Value>,
-    ) -> bool {
-        if let Some(existing_index) = self.tool_indexes.get(tool_id).copied() {
-            if let Some(existing) = self.tool_blocks.get_mut(&existing_index) {
-                if existing.input.is_none() {
-                    existing.input = input;
-                }
-            }
-            return false;
-        }
-
-        self.tool_indexes.insert(tool_id.to_string(), index);
-        self.tool_blocks.insert(
-            index,
-            ToolUseState {
-                id: tool_id.to_string(),
-                name: tool_name.to_string(),
-                input_json: String::new(),
-                input,
-            },
-        );
-        true
-    }
-
-    fn append_input_json(&mut self, index: usize, partial_json: &str) {
-        if let Some(tool) = self.tool_blocks.get_mut(&index) {
-            tool.input_json.push_str(partial_json);
-        }
-    }
-
-    fn tool_by_id(&self, tool_use_id: &str) -> Option<&ToolUseState> {
-        let index = self.tool_indexes.get(tool_use_id)?;
-        self.tool_blocks.get(index)
-    }
-}
-
-#[derive(Debug)]
-struct TurnInProgressGuard {
-    flag: Arc<AtomicBool>,
-    armed: bool,
-}
-
-impl TurnInProgressGuard {
-    fn new(flag: Arc<AtomicBool>) -> Self {
-        Self { flag, armed: true }
-    }
-
-    fn disarm(&mut self) {
-        self.armed = false;
-    }
-}
-
-impl Drop for TurnInProgressGuard {
-    fn drop(&mut self) {
-        if self.armed {
-            self.flag.store(false, Ordering::SeqCst);
-        }
-    }
-}
-
 impl ClaudeHarness {
     pub fn new(events: broadcast::Sender<SessionEvent>) -> Self {
         Self {
@@ -245,274 +79,6 @@ impl ClaudeHarness {
             stderr_task: None,
             shutdown_requested: Arc::new(AtomicBool::new(false)),
         }
-    }
-
-    /// Parse a single NDJSON line and emit SessionEvents.
-    fn process_line(
-        line: &str,
-        turn_id: &str,
-        events: &broadcast::Sender<SessionEvent>,
-        state: &mut ReaderState,
-    ) -> bool {
-        let Ok(value) = serde_json::from_str::<Value>(line) else {
-            return false;
-        };
-        let event = value
-            .get("stream_event")
-            .and_then(|stream| stream.get("event"))
-            .unwrap_or(&value);
-
-        let Some(event_type) = event.get("type").and_then(Value::as_str) else {
-            return false;
-        };
-
-        match event_type {
-            "system" => {
-                if let Some(session_id) = event
-                    .get("session_id")
-                    .or_else(|| value.get("session_id"))
-                    .and_then(Value::as_str)
-                {
-                    let _ = events.send(SessionEvent::ProviderSessionId {
-                        provider_session_id: session_id.to_string(),
-                    });
-                }
-            }
-
-            "content_block_start" => {
-                let Some(index) = event.get("index").and_then(Value::as_u64) else {
-                    return false;
-                };
-                let Some(block) = event.get("content_block") else {
-                    return false;
-                };
-                let Some(block_type) = block.get("type").and_then(Value::as_str) else {
-                    return false;
-                };
-                let index = index as usize;
-
-                if block_type == "tool_use" {
-                    let Some(tool_id) = block.get("id").and_then(Value::as_str) else {
-                        return false;
-                    };
-                    let Some(tool_name) = block.get("name").and_then(Value::as_str) else {
-                        return false;
-                    };
-                    if state.track_tool(index, tool_id, tool_name, None) {
-                        let item = infer_item(tool_name, tool_id, None);
-                        let _ = events.send(SessionEvent::ItemStarted {
-                            turn_id: turn_id.to_string(),
-                            item,
-                        });
-                    }
-                }
-            }
-
-            "content_block_delta" => {
-                let index = event
-                    .get("index")
-                    .and_then(Value::as_u64)
-                    .map(|v| v as usize);
-                let Some(delta) = event.get("delta") else {
-                    return false;
-                };
-                let Some(delta_type) = delta.get("type").and_then(Value::as_str) else {
-                    return false;
-                };
-
-                match delta_type {
-                    "text_delta" => {
-                        if let Some(text) = delta.get("text").and_then(Value::as_str) {
-                            let _ = events.send(SessionEvent::TextDelta {
-                                turn_id: turn_id.to_string(),
-                                content: text.to_string(),
-                            });
-                        }
-                    }
-                    "thinking_delta" | "summary_delta" => {
-                        if let Some(text) = delta
-                            .get("thinking")
-                            .or_else(|| delta.get("summary"))
-                            .and_then(Value::as_str)
-                        {
-                            let _ = events.send(SessionEvent::ReasoningDelta {
-                                turn_id: turn_id.to_string(),
-                                content: text.to_string(),
-                            });
-                        }
-                    }
-                    "input_json_delta" => {
-                        if let (Some(idx), Some(json_chunk)) =
-                            (index, delta.get("partial_json").and_then(Value::as_str))
-                        {
-                            state.append_input_json(idx, json_chunk);
-                        }
-                    }
-                    _ => {}
-                }
-            }
-
-            "content_block_stop" => {
-                // Tool blocks will get completed when we see the tool result.
-                // Text blocks don't need explicit completion.
-            }
-
-            "result" => {
-                let status = if event
-                    .get("is_error")
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false)
-                {
-                    TurnStatus::Failed
-                } else {
-                    TurnStatus::Completed
-                };
-                let _ = events.send(SessionEvent::TurnCompleted {
-                    turn_id: turn_id.to_string(),
-                    status,
-                });
-                return true;
-            }
-
-            // Claude emits "assistant" events with full message content.
-            // These contain tool_use blocks and text blocks.
-            "assistant" => {
-                Self::process_assistant_message(event, turn_id, events, state);
-            }
-
-            // Tool result events: type "user" with tool_result content.
-            "user" => {
-                Self::process_user_message(event, turn_id, events, state);
-            }
-
-            _ => {}
-        }
-
-        false
-    }
-
-    fn process_assistant_message(
-        value: &Value,
-        turn_id: &str,
-        events: &broadcast::Sender<SessionEvent>,
-        state: &mut ReaderState,
-    ) {
-        let Some(blocks) = Self::message_content(value) else {
-            return;
-        };
-
-        for (idx, block) in blocks.iter().enumerate() {
-            let Some(block_type) = block.get("type").and_then(Value::as_str) else {
-                continue;
-            };
-            match block_type {
-                "tool_use" => {
-                    let tool_id = block
-                        .get("id")
-                        .and_then(Value::as_str)
-                        .unwrap_or_default()
-                        .to_string();
-                    let tool_name = block
-                        .get("name")
-                        .and_then(Value::as_str)
-                        .unwrap_or_default()
-                        .to_string();
-                    let input = block.get("input").cloned();
-
-                    // If we haven't already emitted ItemStarted via streaming,
-                    // emit both started and track it.
-                    if state.track_tool(idx, &tool_id, &tool_name, input.clone()) {
-                        let item = infer_item(&tool_name, &tool_id, input);
-                        let _ = events.send(SessionEvent::ItemStarted {
-                            turn_id: turn_id.to_string(),
-                            item,
-                        });
-                    }
-                }
-                "text" => {
-                    if let Some(text) = block.get("text").and_then(Value::as_str) {
-                        if !text.is_empty() {
-                            let _ = events.send(SessionEvent::TextDelta {
-                                turn_id: turn_id.to_string(),
-                                content: text.to_string(),
-                            });
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn process_user_message(
-        value: &Value,
-        turn_id: &str,
-        events: &broadcast::Sender<SessionEvent>,
-        state: &mut ReaderState,
-    ) {
-        let Some(blocks) = Self::message_content(value) else {
-            return;
-        };
-
-        for block in blocks {
-            let Some(block_type) = block.get("type").and_then(Value::as_str) else {
-                continue;
-            };
-            if block_type != "tool_result" {
-                continue;
-            }
-
-            let tool_use_id = block
-                .get("tool_use_id")
-                .and_then(Value::as_str)
-                .unwrap_or_default();
-
-            let Some(tool) = state.tool_by_id(tool_use_id) else {
-                continue;
-            };
-
-            // Extract output from content array or text.
-            let output = Self::extract_tool_result_text(block);
-            let input = tool.parsed_input();
-            let completed_item =
-                build_item(&tool.name, &tool.id, input, ItemStatus::Completed, output);
-            let _ = events.send(SessionEvent::ItemCompleted {
-                turn_id: turn_id.to_string(),
-                item: completed_item,
-            });
-        }
-    }
-
-    fn message_content(value: &Value) -> Option<&[Value]> {
-        value
-            .get("message")
-            .and_then(|m| m.get("content"))
-            .and_then(Value::as_array)
-            .map(Vec::as_slice)
-    }
-
-    fn extract_tool_result_text(block: &Value) -> Option<String> {
-        // Try nested content array first.
-        if let Some(content) = block.get("content").and_then(Value::as_array) {
-            let texts: Vec<&str> = content
-                .iter()
-                .filter_map(|c| {
-                    if c.get("type").and_then(Value::as_str) == Some("text") {
-                        c.get("text").and_then(Value::as_str)
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-            if !texts.is_empty() {
-                return Some(texts.join("\n"));
-            }
-        }
-        // Try direct text field.
-        block
-            .get("content")
-            .and_then(Value::as_str)
-            .map(String::from)
     }
 }
 
@@ -620,7 +186,7 @@ impl SessionHarness for ClaudeHarness {
                     continue;
                 }
 
-                if ClaudeHarness::process_line(&line, &reader_turn_id, &events, &mut state) {
+                if claude_mapping::process_line(&line, &reader_turn_id, &events, &mut state) {
                     saw_turn_completed = true;
                     break;
                 }
@@ -640,16 +206,7 @@ impl SessionHarness for ClaudeHarness {
             turn_in_progress.store(false, Ordering::SeqCst);
         }));
 
-        // Spawn stderr task for logging.
-        self.stderr_task = Some(tokio::spawn(async move {
-            let reader = BufReader::new(stderr);
-            let mut lines = reader.lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                if !line.trim().is_empty() {
-                    tracing::debug!(target: "claude_harness", "{line}");
-                }
-            }
-        }));
+        self.stderr_task = Some(spawn_stderr_logger(stderr, "claude_harness"));
 
         turn_guard.disarm();
         Ok(())
@@ -682,76 +239,6 @@ impl SessionHarness for ClaudeHarness {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn infer_item_bash_to_command() {
-        let item = infer_item("Bash", "tu_1", Some(json!({"command": "cargo test"})));
-        match item {
-            SessionItem::Command { id, command, .. } => {
-                assert_eq!(id, "tu_1");
-                assert_eq!(command, vec!["cargo test"]);
-            }
-            other => panic!("expected Command, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn infer_item_edit_to_file() {
-        let item = infer_item("Edit", "tu_2", Some(json!({"file_path": "src/main.rs"})));
-        match item {
-            SessionItem::File { id, changes, .. } => {
-                assert_eq!(id, "tu_2");
-                assert_eq!(changes[0].path, "src/main.rs");
-                assert_eq!(changes[0].kind.as_deref(), Some("edit"));
-            }
-            other => panic!("expected File, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn infer_item_write_to_file() {
-        let item = infer_item("Write", "tu_3", Some(json!({"file_path": "new.txt"})));
-        match item {
-            SessionItem::File { id, changes, .. } => {
-                assert_eq!(id, "tu_3");
-                assert_eq!(changes[0].path, "new.txt");
-                assert_eq!(changes[0].kind.as_deref(), Some("write"));
-            }
-            other => panic!("expected File, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn infer_item_notebook_edit_to_file() {
-        let item = infer_item(
-            "NotebookEdit",
-            "tu_4",
-            Some(json!({"notebook_path": "analysis.ipynb"})),
-        );
-        match item {
-            SessionItem::File { id, changes, .. } => {
-                assert_eq!(id, "tu_4");
-                assert_eq!(changes[0].path, "analysis.ipynb");
-            }
-            other => panic!("expected File, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn infer_item_unknown_to_tool() {
-        let item = infer_item("WebSearch", "tu_5", Some(json!({"query": "rust async"})));
-        match item {
-            SessionItem::Tool {
-                id, name, input, ..
-            } => {
-                assert_eq!(id, "tu_5");
-                assert_eq!(name, "WebSearch");
-                assert_eq!(input, Some(json!({"query": "rust async"})));
-            }
-            other => panic!("expected Tool, got {other:?}"),
-        }
-    }
 
     #[test]
     fn build_args_minimal() {
@@ -798,168 +285,6 @@ mod tests {
         assert!(args.contains(&"sess_123".to_string()));
         assert!(!args.contains(&"--model".to_string()));
         assert!(!args.contains(&"--dangerously-skip-permissions".to_string()));
-    }
-
-    #[test]
-    fn process_line_system_event() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"system","session_id":"sess_abc123","tools":[]}"#;
-
-        let result = ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-        assert!(!result);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::ProviderSessionId {
-                provider_session_id,
-            } => {
-                assert_eq!(provider_session_id, "sess_abc123");
-            }
-            other => panic!("expected ProviderSessionId event, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_wrapped_stream_event() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"stream_event":{"event":{"type":"system","session_id":"sess_wrapped"}}}"#;
-
-        let result = ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-        assert!(!result);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::ProviderSessionId {
-                provider_session_id,
-            } => {
-                assert_eq!(provider_session_id, "sess_wrapped");
-            }
-            other => panic!("expected ProviderSessionId event, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_text_delta() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"content_block_delta","index":0,"delta":{"type":"text_delta","text":"Hello world"}}"#;
-
-        ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::TextDelta { turn_id, content } => {
-                assert_eq!(turn_id, "turn_1");
-                assert_eq!(content, "Hello world");
-            }
-            other => panic!("expected TextDelta, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_thinking_delta() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"content_block_delta","index":0,"delta":{"type":"thinking_delta","thinking":"Let me think..."}}"#;
-
-        ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::ReasoningDelta { turn_id, content } => {
-                assert_eq!(turn_id, "turn_1");
-                assert_eq!(content, "Let me think...");
-            }
-            other => panic!("expected ReasoningDelta, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_result_completes_turn() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"result","duration_ms":1234,"cost_usd":0.01,"is_error":false,"result":"done","session_id":"sess_abc"}"#;
-
-        let result = ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-        assert!(result);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::TurnCompleted { turn_id, status } => {
-                assert_eq!(turn_id, "turn_1");
-                assert_eq!(status, TurnStatus::Completed);
-            }
-            other => panic!("expected TurnCompleted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_result_error_marks_failed() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"result","is_error":true,"result":"failed"}"#;
-
-        let result = ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-        assert!(result);
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::TurnCompleted { turn_id, status } => {
-                assert_eq!(turn_id, "turn_1");
-                assert_eq!(status, TurnStatus::Failed);
-            }
-            other => panic!("expected TurnCompleted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_tool_use_start() {
-        let (tx, mut rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-        let line = r#"{"type":"content_block_start","index":1,"content_block":{"type":"tool_use","id":"tu_bash_1","name":"Bash"}}"#;
-
-        ClaudeHarness::process_line(line, "turn_1", &tx, &mut state);
-
-        assert!(state.tool_blocks.contains_key(&1));
-        let tool = &state.tool_blocks[&1];
-        assert_eq!(tool.id, "tu_bash_1");
-        assert_eq!(tool.name, "Bash");
-        assert_eq!(state.tool_indexes.get("tu_bash_1"), Some(&1));
-
-        let event = rx.try_recv().expect("should have event");
-        match event {
-            SessionEvent::ItemStarted { turn_id, item } => {
-                assert_eq!(turn_id, "turn_1");
-                assert!(matches!(item, SessionItem::Command { .. }));
-            }
-            other => panic!("expected ItemStarted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn process_line_input_json_delta_accumulates() {
-        let (tx, _rx) = broadcast::channel(16);
-        let mut state = ReaderState::default();
-
-        // Start a tool block first.
-        let start_line = r#"{"type":"content_block_start","index":0,"content_block":{"type":"tool_use","id":"tu_1","name":"Bash"}}"#;
-        ClaudeHarness::process_line(start_line, "turn_1", &tx, &mut state);
-
-        // Send partial input.
-        let delta1 = r#"{"type":"content_block_delta","index":0,"delta":{"type":"input_json_delta","partial_json":"{\"com"}}"#;
-        let delta2 = r#"{"type":"content_block_delta","index":0,"delta":{"type":"input_json_delta","partial_json":"mand\":\"ls\"}"}}"#;
-        ClaudeHarness::process_line(delta1, "turn_1", &tx, &mut state);
-        ClaudeHarness::process_line(delta2, "turn_1", &tx, &mut state);
-
-        assert_eq!(
-            state
-                .tool_blocks
-                .get(&0)
-                .map(|tool| tool.input_json.as_str()),
-            Some("{\"command\":\"ls\"}")
-        );
     }
 
     #[tokio::test]
