+++ rust/loopflow/src/lfd/sessions/harness/codex.rs
@@ -10,10 +10,10 @@ use tokio::process::{Child, Command};
 use tokio::sync::{broadcast, mpsc, oneshot, Mutex};
 use tokio::task::JoinHandle;
 
-use crate::lfd::sessions::harness::SessionHarness;
-use crate::lfd::sessions::types::{
-    FileEdit, ItemDelta, ItemStatus, SessionConfig, SessionEvent, SessionItem, TurnStatus,
-};
+use crate::lfd::sessions::harness::codex_mapping::ItemPhase;
+use crate::lfd::sessions::harness::common::spawn_stderr_logger;
+use crate::lfd::sessions::harness::{codex_mapping, SessionHarness};
+use crate::lfd::sessions::types::{SessionConfig, SessionEvent};
 
 #[derive(Debug)]
 enum OutboundRpc {
@@ -26,12 +26,6 @@ enum OutboundRpc {
         id: Value,
         result: Value,
     },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ItemPhase {
-    Started,
-    Completed,
 }
 
 pub struct CodexHarness {
@@ -84,194 +78,6 @@ impl CodexHarness {
         .map_err(|_| anyhow!("codex writer task unavailable"))
     }
 
-    fn extract_turn_id(params: &Value) -> Option<String> {
-        params
-            .get("turn")
-            .and_then(|t| t.get("id"))
-            .and_then(Value::as_str)
-            .or_else(|| params.get("turnId").and_then(Value::as_str))
-            .map(ToString::to_string)
-    }
-
-    fn item_payload(params: &Value) -> &Value {
-        params.get("item").unwrap_or(params)
-    }
-
-    fn text_field(value: &Value, key: &str) -> Option<String> {
-        value
-            .get(key)
-            .and_then(Value::as_str)
-            .map(ToString::to_string)
-    }
-
-    fn required_text_field(value: &Value, key: &str) -> String {
-        Self::text_field(value, key).unwrap_or_default()
-    }
-
-    fn map_item_id(params: &Value) -> String {
-        Self::item_payload(params)
-            .get("id")
-            .and_then(Value::as_str)
-            .or_else(|| params.get("itemId").and_then(Value::as_str))
-            .or_else(|| params.get("id").and_then(Value::as_str))
-            .unwrap_or("unknown")
-            .to_string()
-    }
-
-    fn map_item_type(params: &Value) -> &str {
-        Self::item_payload(params)
-            .get("type")
-            .and_then(Value::as_str)
-            .or_else(|| params.get("kind").and_then(Value::as_str))
-            .unwrap_or("tool")
-    }
-
-    fn map_turn_status(params: &Value) -> TurnStatus {
-        let turn_status = params
-            .get("turn")
-            .and_then(|t| t.get("status"))
-            .and_then(Value::as_str)
-            .or_else(|| params.get("status").and_then(Value::as_str))
-            .unwrap_or_default();
-        match turn_status {
-            "interrupted" | "cancelled" => TurnStatus::Interrupted,
-            "failed" | "error" => TurnStatus::Failed,
-            _ => TurnStatus::Completed,
-        }
-    }
-
-    fn map_item_status(params: &Value) -> ItemStatus {
-        let status = Self::item_payload(params)
-            .get("status")
-            .and_then(Value::as_str)
-            .or_else(|| params.get("status").and_then(Value::as_str))
-            .unwrap_or("in_progress");
-        match status {
-            "completed" => ItemStatus::Completed,
-            "failed" | "error" => ItemStatus::Failed,
-            "declined" => ItemStatus::Declined,
-            _ => ItemStatus::InProgress,
-        }
-    }
-
-    fn parse_command(item: &Value) -> Vec<String> {
-        item.get("command")
-            .and_then(Value::as_array)
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(Value::as_str)
-                    .map(ToString::to_string)
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    fn build_item(params: &Value, phase: ItemPhase) -> SessionItem {
-        let id = Self::map_item_id(params);
-        let item_type = Self::map_item_type(params);
-        let item = Self::item_payload(params);
-        let status = Self::map_item_status(params);
-        let completed = phase == ItemPhase::Completed;
-
-        match item_type {
-            "commandExecution" => SessionItem::Command {
-                id,
-                command: Self::parse_command(item),
-                cwd: Self::required_text_field(item, "cwd"),
-                status,
-                output: if completed {
-                    Self::text_field(item, "aggregatedOutput")
-                } else {
-                    None
-                },
-                exit_code: if completed {
-                    item.get("exitCode")
-                        .and_then(Value::as_i64)
-                        .map(|v| v as i32)
-                } else {
-                    None
-                },
-                duration_ms: if completed {
-                    item.get("durationMs").and_then(Value::as_u64)
-                } else {
-                    None
-                },
-            },
-            "fileChange" => SessionItem::File {
-                id,
-                changes: parse_file_changes(item),
-                status,
-            },
-            "mcpToolCall" => SessionItem::Tool {
-                id,
-                name: Self::mcp_tool_name(item),
-                status,
-                input: Some(item.get("arguments").cloned().unwrap_or_else(|| json!({}))),
-                output: if completed {
-                    Self::text_field(item, "result").or_else(|| Self::text_field(item, "error"))
-                } else {
-                    None
-                },
-            },
-            "agentMessage" => SessionItem::Message {
-                id,
-                text: Self::required_text_field(item, "text"),
-                phase: Self::text_field(item, "phase"),
-            },
-            "plan" => SessionItem::Thought {
-                id,
-                text: Self::required_text_field(item, "text"),
-            },
-            _ => SessionItem::Tool {
-                id,
-                name: item_type.to_string(),
-                status,
-                input: item.get("input").cloned(),
-                output: if completed {
-                    Self::text_field(item, "output")
-                } else {
-                    None
-                },
-            },
-        }
-    }
-
-    fn mcp_tool_name(item: &Value) -> String {
-        let tool = Self::required_text_field(item, "tool");
-        let server = Self::text_field(item, "server").unwrap_or_default();
-        if server.is_empty() {
-            if tool.is_empty() {
-                "mcp_tool_call".to_string()
-            } else {
-                tool
-            }
-        } else if tool.is_empty() {
-            server
-        } else {
-            format!("{server}/{tool}")
-        }
-    }
-
-    fn map_item_delta(method: &str, params: &Value) -> Option<ItemDelta> {
-        let content = Self::text_content(params)?;
-        match method {
-            "item/commandExecution/outputDelta" | "item/fileChange/outputDelta" => {
-                Some(ItemDelta::Output { content })
-            }
-            "item/plan/delta" => Some(ItemDelta::PlanText { content }),
-            _ => None,
-        }
-    }
-
-    fn text_content(params: &Value) -> Option<String> {
-        params
-            .get("content")
-            .and_then(Value::as_str)
-            .or_else(|| params.get("delta").and_then(Value::as_str))
-            .or_else(|| params.get("output").and_then(Value::as_str))
-            .map(ToString::to_string)
-    }
-
     async fn resolve_turn_id(
         turn_id_from_params: Option<&str>,
         current_turn_id: &Arc<Mutex<Option<String>>>,
@@ -301,31 +107,6 @@ impl CodexHarness {
             let _ = handle.await;
         }
     }
-}
-
-fn parse_file_changes(item: &Value) -> Vec<FileEdit> {
-    item.get("changes")
-        .and_then(Value::as_array)
-        .map(|arr| {
-            arr.iter()
-                .map(|c| FileEdit {
-                    path: c
-                        .get("path")
-                        .and_then(Value::as_str)
-                        .unwrap_or_default()
-                        .to_string(),
-                    kind: c
-                        .get("kind")
-                        .and_then(Value::as_str)
-                        .map(ToString::to_string),
-                    diff: c
-                        .get("diff")
-                        .and_then(Value::as_str)
-                        .map(ToString::to_string),
-                })
-                .collect()
-        })
-        .unwrap_or_default()
 }
 
 #[async_trait]
@@ -474,7 +255,7 @@ impl CodexHarness {
                 }
 
                 // Resolve turn_id from notification params or tracked state.
-                let turn_id_from_params = CodexHarness::extract_turn_id(&params);
+                let turn_id_from_params = codex_mapping::extract_turn_id(&params);
 
                 match method {
                     "turn/started" => {
@@ -493,7 +274,7 @@ impl CodexHarness {
                         .await;
                         turn_in_progress.store(false, Ordering::Relaxed);
                         *current_turn_id.lock().await = None;
-                        let status = CodexHarness::map_turn_status(&params);
+                        let status = codex_mapping::map_turn_status(&params);
                         let _ = event_tx.send(SessionEvent::TurnCompleted {
                             turn_id: tid,
                             status,
@@ -505,7 +286,7 @@ impl CodexHarness {
                             &current_turn_id,
                         )
                         .await;
-                        let item = CodexHarness::build_item(&params, ItemPhase::Started);
+                        let item = codex_mapping::build_item(&params, ItemPhase::Started);
                         let _ = event_tx.send(SessionEvent::ItemStarted { turn_id: tid, item });
                     }
                     "item/completed" => {
@@ -514,11 +295,11 @@ impl CodexHarness {
                             &current_turn_id,
                         )
                         .await;
-                        let item = CodexHarness::build_item(&params, ItemPhase::Completed);
+                        let item = codex_mapping::build_item(&params, ItemPhase::Completed);
                         let _ = event_tx.send(SessionEvent::ItemCompleted { turn_id: tid, item });
                     }
                     "item/agentMessage/delta" => {
-                        if let Some(content) = CodexHarness::text_content(&params) {
+                        if let Some(content) = codex_mapping::text_content(&params) {
                             let tid = CodexHarness::resolve_turn_id(
                                 turn_id_from_params.as_deref(),
                                 &current_turn_id,
@@ -534,7 +315,7 @@ impl CodexHarness {
                         // Final agent message text is captured in item/completed.
                     }
                     "item/reasoning/summaryTextDelta" | "item/reasoning/textDelta" => {
-                        if let Some(content) = CodexHarness::text_content(&params) {
+                        if let Some(content) = codex_mapping::text_content(&params) {
                             let tid = CodexHarness::resolve_turn_id(
                                 turn_id_from_params.as_deref(),
                                 &current_turn_id,
@@ -549,13 +330,13 @@ impl CodexHarness {
                     "item/commandExecution/outputDelta"
                     | "item/fileChange/outputDelta"
                     | "item/plan/delta" => {
-                        if let Some(data) = CodexHarness::map_item_delta(method, &params) {
+                        if let Some(data) = codex_mapping::map_item_delta(method, &params) {
                             let tid = CodexHarness::resolve_turn_id(
                                 turn_id_from_params.as_deref(),
                                 &current_turn_id,
                             )
                             .await;
-                            let item_id = CodexHarness::map_item_id(&params);
+                            let item_id = codex_mapping::map_item_id(&params);
                             let _ = event_tx.send(SessionEvent::ItemUpdated {
                                 turn_id: tid,
                                 item_id,
@@ -606,12 +387,7 @@ impl CodexHarness {
             }
         });
 
-        let stderr_task = tokio::spawn(async move {
-            let mut lines = BufReader::new(stderr).lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                tracing::debug!(target: "lfd::sessions::codex", stderr = %line, "codex stderr");
-            }
-        });
+        let stderr_task = spawn_stderr_logger(stderr, "lfd::sessions::codex");
 
         self.child = Some(child);
         self.outbound_tx = Some(outbound_tx);
@@ -636,114 +412,5 @@ impl CodexHarness {
             .await?;
 
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn build_item_maps_file_change_to_file() {
-        let params = json!({
-            "item": {
-                "id": "item_1",
-                "type": "fileChange",
-                "status": "completed",
-                "changes": [
-                    {"path": "src/main.rs", "kind": "update", "diff": "-a\n+b"}
-                ]
-            }
-        });
-
-        let item = CodexHarness::build_item(&params, ItemPhase::Completed);
-        match item {
-            SessionItem::File {
-                id,
-                changes,
-                status,
-            } => {
-                assert_eq!(id, "item_1");
-                assert_eq!(status, ItemStatus::Completed);
-                assert_eq!(changes.len(), 1);
-                assert_eq!(changes[0].path, "src/main.rs");
-            }
-            other => panic!("expected file item, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn build_item_maps_agent_message_to_message() {
-        let params = json!({
-            "item": {
-                "id": "item_2",
-                "type": "agentMessage",
-                "text": "Done",
-                "phase": "final"
-            }
-        });
-
-        let item = CodexHarness::build_item(&params, ItemPhase::Completed);
-        match item {
-            SessionItem::Message { id, text, phase } => {
-                assert_eq!(id, "item_2");
-                assert_eq!(text, "Done");
-                assert_eq!(phase.as_deref(), Some("final"));
-            }
-            other => panic!("expected message item, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn build_item_maps_plan_to_thought() {
-        let params = json!({
-            "item": {
-                "id": "item_3",
-                "type": "plan",
-                "text": "Run tests first"
-            }
-        });
-
-        let item = CodexHarness::build_item(&params, ItemPhase::Completed);
-        match item {
-            SessionItem::Thought { id, text } => {
-                assert_eq!(id, "item_3");
-                assert_eq!(text, "Run tests first");
-            }
-            other => panic!("expected thought item, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn build_item_maps_mcp_tool_call_to_generic_tool() {
-        let params = json!({
-            "item": {
-                "id": "item_4",
-                "type": "mcpToolCall",
-                "status": "completed",
-                "server": "github",
-                "tool": "search",
-                "arguments": { "query": "regression" },
-                "result": "ok"
-            }
-        });
-
-        let item = CodexHarness::build_item(&params, ItemPhase::Completed);
-        match item {
-            SessionItem::Tool {
-                id,
-                name,
-                status,
-                input,
-                output,
-            } => {
-                assert_eq!(id, "item_4");
-                assert_eq!(name, "github/search");
-                assert_eq!(status, ItemStatus::Completed);
-                assert_eq!(input, Some(json!({ "query": "regression" })));
-                assert_eq!(output.as_deref(), Some("ok"));
-            }
-            other => panic!("expected tool item, got {other:?}"),
-        }
     }
 }
